fix(frontend): 修复封禁时间戳显示为 1970 年的整数溢出 bug
timestamp as u32 * 1000 在 Unix 时间戳超过 u32 范围时溢出回绕，
改为 as f64 * 1000.0 避免溢出。

+++ frontend/src/format/mod.rs
@@ -47,7 +47,8 @@ pub fn format_datetime(timestamp: i64) -> String {
     if timestamp <= 0 {
         return "N/A".to_string();
     }
-    let date = js_sys::Date::new(&(timestamp as u32 * 1000).into());
+    // SAFETY: js_sys::Date 接受毫秒时间戳，i64 转 f64 在合理范围内精度足够
+    let date = js_sys::Date::new(&(timestamp as f64 * 1000.0).into());
     let year = date.get_full_year();
     let month = date.get_month() + 1;
     let day = date.get_date();
